Handle fetching from DB before entry created
When we start a journey we:
- create an entry for that journey in the database
- redirect to a page which then looks up that entry from the database in
  order to render the page

The order in which this operations complete is not guaranteed to be the
same. So we were seeing app crash when it was trying to get data that
the database didn’t yet know about.

Not having the data isn’t actually important for the first page, because
we don’t play any of it back on the page.